requirements.txt
aiohttp==3.7.4
altgraph==0.17
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.3
certifi==2022.12.7
cffi==1.14.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.3.2
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.1.14
django-cleanup==5.1.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
djangorestframework==3.11.2
future==0.18.2
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.9.1
mccabe==0.6.1
multidict==5.0.2
oauthlib==3.1.0
pefile==2019.4.18
pillow>=8.3.2
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyinstaller-hooks-contrib==2020.11
pylint==2.5.3
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.1.0
boto3==1.18.1
django-storages==1.11.1
botocore==1.21.1